requirements.txt
nedrex==0.1.4
tqdm
pyyaml
jupyter
pandas